// api/src/datatypes/entities.d.ts
/* tslint:disable */
import {HKNArea, InterviewOutcome, ManRole, SlotState} from "./enums";
import {Change, TimeSlot} from "./dataTypes";
import {Person} from "./person";
import {Application} from "./application";

/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Member = Person & {
  id: string;
  role: ManRole;
  is_expert: boolean;
  is_board: boolean;
  [k: string]: unknown;
};
export type Applicant = Person & {
  id: string;
  birth_date: string;
  telegram_uid?: string;
  how_know_HKN: string;
  [k: string]: unknown;
};
export type BscApplication = Application & {
  study_path: string;
  academic_year: number;
  cfu: number;
  grades?: string;
  grades_avg: number;
  [k: string]: unknown;
};
export type MscApplication = Application & {
  [k: string]: unknown;
};
export type PhdApplication = Application & {
  msc_study_path: string;
  phd_description: string;
  [k: string]: unknown;
};
export interface Slot {
  id: number;
  state: SlotState;
  cal_id?: string;
  time_slot: TimeSlot;
  members: [
    {
      id: string;
      name: string;
      surname: string;
      sex: "male" | "female";
      image: string;
      is_board: boolean;
      is_expert: boolean;
      [k: string]: unknown;
    },
    {
      id: string;
      name: string;
      surname: string;
      sex: "male" | "female";
      image: string;
      is_board: boolean;
      is_expert: boolean;
      [k: string]: unknown;
    },
    ...{
      id: string;
      name: string;
      surname: string;
      sex: "male" | "female";
      image: string;
      is_board: boolean;
      is_expert: boolean;
      [k: string]: unknown;
    }[]
  ];
  interview_id: number;
  [k: string]: unknown;
}
export interface Interview {
  outcome?: InterviewOutcome;
  last_modified: Change;
  finalized_at?: string;
  assigned_area?:HKNArea;
  notes: string;
  application_id: number;
  [k: string]: unknown;
}